feat(tasks): implementar RBAC básico — ocultar acciones según permisos del usuario
